Introduce dma_access_mode enum
In preparation for adding an alternative way of message-based DMA,
introduce an enum to make the access mode explicit. This is needed to
distinguish mmap-based DMA access from message based over a separate
socket. Both require a file descriptor to be included with the
DMA mapping request, and so the presence of the file descriptor is no
longer sufficient to indicate how to perform DMA accesses.