Have client report page request completion
This changes the DMA page request flow to require the client to send an
explicit completion once the request has been processed. The completion
status is then reported back to server code via a callback. This fixes
page requests for the case where merely waiting for a DMA map callback
doesn't work correctly, such as when multiple mappings are required to
satisfy the request or the request fails / is rejected.

On the protocol level, page request completion is indicated to the
server by a VFIO_USER_DMA_PAGE_REQUEST message from client to server.
This is not how this would be done in a production implementation, the
protocol would rather use an asynchronous reply to the server-to-client
VFIO_USER_DMA_PAGE_REQUEST command. Unfortunately, the current state of
the transaction layer doesn't allow asynchronous replies to work
properly, so we're taking a short cut here to keep this proof-of-concept
page request implementation simple.